workflow: split actor-driven mechanics
