+++ packages/core/src/sprints/observe/sources.py
@@ -29,6 +29,11 @@ from sprints.engine.state import (
 from sprints.engine.work import work_item_from_issue
 from sprints.core.contracts import WorkflowContractError, load_workflow_contract
 from sprints.core.paths import runtime_paths
+from sprints.workflows.projection import (
+    project_lane_map,
+    project_state_lane,
+    projected_lane_is_terminal,
+)
 
 
 def _read_jsonl_tail(path: Path, limit: int) -> list[dict[str, Any]]:
@@ -154,131 +159,7 @@ def _state_lanes_by_id(
 
 
 def _lane_is_terminal(lane: dict[str, Any]) -> bool:
-    return str(lane.get("status") or "").strip() in {
-        "complete",
-        "released",
-        "merged",
-        "closed",
-        "archived",
-    }
-
-
-def _state_lane_entry(lane: dict[str, Any], *, workflow_name: str) -> dict[str, Any]:
-    issue = lane.get("issue") if isinstance(lane.get("issue"), dict) else {}
-    lane_id = str(lane.get("lane_id") or issue.get("id") or "").strip()
-    identifier = str(
-        issue.get("identifier") or issue.get("number") or lane_id or "unknown"
-    )
-    status = str(lane.get("status") or "active").strip() or "active"
-    stage = str(lane.get("stage") or status).strip() or status
-    work_item = work_item_from_issue(
-        {
-            "id": issue.get("id") or lane_id or identifier,
-            "identifier": identifier,
-            "title": issue.get("title") or "",
-            "url": issue.get("url"),
-            "state": status,
-        },
-        source=workflow_name,
-    ).to_dict()
-    pull_request = (
-        lane.get("pull_request") if isinstance(lane.get("pull_request"), dict) else {}
-    )
-    pending_retry = (
-        lane.get("pending_retry") if isinstance(lane.get("pending_retry"), dict) else {}
-    )
-    retry_history = [
-        record for record in lane.get("retry_history") or [] if isinstance(record, dict)
-    ]
-    retry_latest = retry_history[-1] if retry_history else {}
-    attention = (
-        lane.get("operator_attention")
-        if isinstance(lane.get("operator_attention"), dict)
-        else {}
-    )
-    runtime_session = (
-        lane.get("runtime_session")
-        if isinstance(lane.get("runtime_session"), dict)
-        else {}
-    )
-    actor_dispatch = (
-        lane.get("actor_dispatch")
-        if isinstance(lane.get("actor_dispatch"), dict)
-        else {}
-    )
-    dispatch_runtime = (
-        actor_dispatch.get("runtime")
-        if isinstance(actor_dispatch.get("runtime"), dict)
-        else {}
-    )
-    tracker = lane.get("tracker") if isinstance(lane.get("tracker"), dict) else {}
-    review_signals = (
-        lane.get("review_signals")
-        if isinstance(lane.get("review_signals"), dict)
-        else {}
-    )
-    merge_signal = (
-        lane.get("merge_signal") if isinstance(lane.get("merge_signal"), dict) else {}
-    )
-    return {
-        "lane_id": lane_id or identifier,
-        "state": stage,
-        "workflow_state": stage,
-        "issue_number": issue.get("number"),
-        "issue_identifier": identifier,
-        "issue_title": issue.get("title"),
-        "lane_status": status,
-        "status": status,
-        "stage": stage,
-        "actor": lane.get("actor"),
-        "actor_mode": runtime_session.get("actor_mode")
-        or dispatch_runtime.get("actor_mode")
-        or dispatch_runtime.get("mode"),
-        "attempt": lane.get("attempt"),
-        "board_state": tracker.get("board_state") or lane.get("board_state"),
-        "tracker": tracker or None,
-        "branch": lane.get("branch"),
-        "pull_request": pull_request or None,
-        "pull_request_number": pull_request.get("number"),
-        "pull_request_url": pull_request.get("url"),
-        "review_signals": review_signals or None,
-        "merge_signal": merge_signal or None,
-        "review_required_change_count": len(
-            review_signals.get("required_changes") or []
-        ),
-        "reviewer_actor_running": review_signals.get("reviewer_actor_running"),
-        "merge_signal_seen": merge_signal.get("seen")
-        or review_signals.get("merge_signal_seen"),
-        "retry_at": pending_retry.get("due_at"),
-        "retry_target": pending_retry.get("target"),
-        "retry_attempt": pending_retry.get("attempt"),
-        "retry_current_attempt": pending_retry.get("current_attempt")
-        or retry_latest.get("current_attempt"),
-        "retry_max_attempts": pending_retry.get("max_attempts")
-        or retry_latest.get("max_attempts"),
-        "retry_delay_seconds": pending_retry.get("delay_seconds")
-        or retry_latest.get("delay_seconds"),
-        "retry_backoff_seconds": pending_retry.get("delay_seconds")
-        or retry_latest.get("delay_seconds"),
-        "retry_reason": pending_retry.get("reason") or retry_latest.get("reason"),
-        "retry_history_count": len(retry_history),
-        "operator_attention_reason": attention.get("reason"),
-        "operator_attention_message": attention.get("message"),
-        "last_progress_at": lane.get("last_progress_at"),
-        "runtime_status": runtime_session.get("status"),
-        "dispatch_id": actor_dispatch.get("dispatch_id"),
-        "dispatch_status": actor_dispatch.get("status"),
-        "dispatch_actor": actor_dispatch.get("actor"),
-        "dispatch_stage": actor_dispatch.get("stage"),
-        "dispatch_mode": dispatch_runtime.get("dispatch_mode"),
-        "dispatch_updated_at": actor_dispatch.get("updated_at"),
-        "dispatch_journal_count": len(lane.get("dispatch_journal") or []),
-        "side_effect_count": len(lane.get("side_effects") or []),
-        "thread_id": lane.get("thread_id") or runtime_session.get("thread_id"),
-        "turn_id": lane.get("turn_id") or runtime_session.get("turn_id"),
-        "kind": status,
-        "work_item": work_item,
-    }
+    return projected_lane_is_terminal(lane)
 
 
 def _engine_work_items(workflow_root: Path, workflow_name: str) -> list[dict[str, Any]]:
@@ -305,115 +186,6 @@ def _engine_runtime_sessions(
         str(session.get("work_id") or session.get("issue_id") or ""): session
         for session in sessions
         if isinstance(session, dict)
-    }
-
-
-def _engine_lane_entry(
-    work_item: dict[str, Any],
-    *,
-    workflow_name: str,
-    state_lane: dict[str, Any] | None,
-    runtime_session: dict[str, Any] | None,
-) -> dict[str, Any]:
-    metadata = (
-        work_item.get("metadata") if isinstance(work_item.get("metadata"), dict) else {}
-    )
-    state_entry = (
-        _state_lane_entry(state_lane, workflow_name=workflow_name)
-        if isinstance(state_lane, dict)
-        else {}
-    )
-    lane_id = str(work_item.get("work_id") or state_entry.get("lane_id") or "")
-    runtime_session = runtime_session if isinstance(runtime_session, dict) else {}
-    pull_request = metadata.get("pull_request") or state_entry.get("pull_request") or {}
-    review_signals = metadata.get("review_signals") or state_entry.get(
-        "review_signals"
-    )
-    merge_signal = metadata.get("merge_signal") or state_entry.get("merge_signal")
-    tracker = metadata.get("tracker") if isinstance(metadata.get("tracker"), dict) else {}
-    attention = (
-        metadata.get("operator_attention")
-        if isinstance(metadata.get("operator_attention"), dict)
-        else {}
-    )
-    if not attention and state_entry.get("operator_attention_reason"):
-        attention = {
-            "reason": state_entry.get("operator_attention_reason"),
-            "message": state_entry.get("operator_attention_message"),
-        }
-    work_item_ref = work_item_from_issue(
-        {
-            "id": lane_id or work_item.get("identifier") or "unknown",
-            "identifier": work_item.get("identifier") or lane_id,
-            "title": work_item.get("title") or "",
-            "url": work_item.get("url"),
-            "state": work_item.get("state"),
-        },
-        source=workflow_name,
-    ).to_dict()
-    return {
-        **state_entry,
-        "lane_id": lane_id,
-        "state": metadata.get("stage")
-        or state_entry.get("state")
-        or work_item.get("state"),
-        "workflow_state": metadata.get("stage")
-        or state_entry.get("workflow_state")
-        or work_item.get("state"),
-        "issue_identifier": work_item.get("identifier")
-        or state_entry.get("issue_identifier")
-        or lane_id,
-        "issue_title": work_item.get("title") or state_entry.get("issue_title"),
-        "lane_status": work_item.get("state") or state_entry.get("lane_status"),
-        "status": work_item.get("state") or state_entry.get("status"),
-        "stage": metadata.get("stage") or state_entry.get("stage"),
-        "actor": metadata.get("actor") or state_entry.get("actor"),
-        "actor_mode": runtime_session.get("actor_mode")
-        or state_entry.get("actor_mode"),
-        "attempt": metadata.get("attempt") or state_entry.get("attempt"),
-        "board_state": tracker.get("board_state") or state_entry.get("board_state"),
-        "tracker": tracker or state_entry.get("tracker"),
-        "branch": metadata.get("branch") or state_entry.get("branch"),
-        "pull_request": pull_request or None,
-        "pull_request_number": (pull_request or {}).get("number")
-        if isinstance(pull_request, dict)
-        else state_entry.get("pull_request_number"),
-        "pull_request_url": (pull_request or {}).get("url")
-        if isinstance(pull_request, dict)
-        else state_entry.get("pull_request_url"),
-        "review_signals": review_signals,
-        "merge_signal": merge_signal,
-        "review_required_change_count": len(
-            (review_signals or {}).get("required_changes") or []
-        )
-        if isinstance(review_signals, dict)
-        else state_entry.get("review_required_change_count"),
-        "reviewer_actor_running": (review_signals or {}).get(
-            "reviewer_actor_running"
-        )
-        if isinstance(review_signals, dict)
-        else state_entry.get("reviewer_actor_running"),
-        "merge_signal_seen": (merge_signal or {}).get("seen")
-        if isinstance(merge_signal, dict)
-        else state_entry.get("merge_signal_seen"),
-        "operator_attention_reason": attention.get("reason"),
-        "operator_attention_message": attention.get("message"),
-        "runtime_status": runtime_session.get("status")
-        or state_entry.get("runtime_status"),
-        "thread_id": runtime_session.get("thread_id")
-        or metadata.get("thread_id")
-        or state_entry.get("thread_id"),
-        "turn_id": runtime_session.get("turn_id")
-        or metadata.get("turn_id")
-        or state_entry.get("turn_id"),
-        "last_progress_at": runtime_session.get("updated_at")
-        or work_item.get("updated_at")
-        or state_entry.get("last_progress_at"),
-        "engine_updated_at": work_item.get("updated_at"),
-        "lane_status_source": "engine_work_items",
-        "state_json_present": isinstance(state_lane, dict),
-        "kind": work_item.get("state") or state_entry.get("kind"),
-        "work_item": work_item_ref,
     }
 
 
@@ -527,31 +299,17 @@ def active_lanes(workflow_root: Path) -> list[dict[str, Any]]:
 
     state_lanes = _state_lanes_by_id(workflow_root, "change-delivery")
     runtime_sessions = _engine_runtime_sessions(workflow_root, "change-delivery")
-    engine_entries: list[dict[str, Any]] = []
-    seen: set[str] = set()
-    for work_item in _engine_work_items(workflow_root, "change-delivery"):
-        if not isinstance(work_item, dict):
-            continue
-        lane_id = str(work_item.get("work_id") or "").strip()
-        if not lane_id:
-            continue
-        entry = _engine_lane_entry(
-            work_item,
-            workflow_name="change-delivery",
-            state_lane=state_lanes.get(lane_id),
-            runtime_session=runtime_sessions.get(lane_id),
-        )
-        seen.add(lane_id)
-        if not _lane_is_terminal({"status": entry.get("status")}):
-            engine_entries.append(entry)
-    for lane in _active_state_lanes(workflow_root, "change-delivery"):
-        lane_id = str(lane.get("lane_id") or "").strip()
-        if lane_id and lane_id not in seen:
-            entry = _state_lane_entry(lane, workflow_name="change-delivery")
-            entry["lane_status_source"] = "workflow_state"
-            entry["state_json_present"] = True
-            engine_entries.append(entry)
-    return engine_entries
+    projected = project_lane_map(
+        workflow_name="change-delivery",
+        state_lanes=state_lanes,
+        engine_work_items=_engine_work_items(workflow_root, "change-delivery"),
+        engine_runtime_sessions=runtime_sessions,
+    )
+    return [
+        lane
+        for lane in projected.values()
+        if not _lane_is_terminal({"status": lane.get("status")})
+    ]
 
 
 def alert_state(workflow_root: Path) -> dict[str, Any]:
@@ -604,7 +362,7 @@ def workflow_status(workflow_root: Path) -> dict[str, Any]:
         latest_runs = _engine_runs(workflow_root, "change-delivery")
         active = _active_state_lanes(workflow_root, "change-delivery")
         active_entries = [
-            _state_lane_entry(lane, workflow_name="change-delivery") for lane in active
+            project_state_lane(lane, workflow_name="change-delivery") for lane in active
         ]
         running_count = len(
             [lane for lane in active if str(lane.get("status") or "") == "running"]
